# [xamlc] add a new ResourceLoader.IsEnabled check