@andreinitescu 2020-08-12
@jonathanpeppers Great stuff, what about calls to the `VisualDiagnostics.RegisterSourceInfo` thing?

![image](https://user-images.githubusercontent.com/743918/89980444-8e50c400-dc7a-11ea-8cf0-46943850e682.png)

What is it for, is it needed?