@jonathanpeppers 2020-08-12
@andreinitescu There might be more things to look into here; this was a first "easy" step.

@StephaneDelcroix depends on what you think, but we could also check for a "Debug" build and skip emitting the `ResourceLoader` checks completely, using:

https://github.com/xamarin/Xamarin.Forms/blob/c6d4d79f35f442d958dba3d8641299f5b506d2fb/Xamarin.Forms.Build.Tasks/XamlCTask.cs#L75

https://github.com/xamarin/Xamarin.Forms/blob/c6d4d79f35f442d958dba3d8641299f5b506d2fb/Xamarin.Forms.Build.Tasks/XamlCTask.cs#L334

We might also want to check the `$(Optimize)` MSBuild property for these, too.