@StephaneDelcroix 2020-08-13
`RegisterSourceInfo` no-op while running without a debugger attached. We could go one step further and avoid the call altogether when building on Release