@StephaneDelcroix 2020-08-13
@jonathanpeppers I now better understand your benchmark on bindings... it's not so much about applying the same binding over and over than creating a page with a binding.