@StephaneDelcroix 2020-08-13
@jonathanpeppers I'd have to check with our good friends of tooling. For the previewer, the check for ResourceProvider had to be in for Release builds (for reasons I no longer recall). For HR I don't think it is. Also, I think your plan is still to skip XamlC for Debug builds, so we might be able to remove this completely