@jonathanpeppers 2020-08-13
The previewer works for Release builds in 16.7 currently:

![image](https://user-images.githubusercontent.com/840039/90136999-9c661980-dd3a-11ea-9cdd-caf096cd5dd4.png)

So I think we have to leave the `ResourceLoader` code in `Release` builds, but we can add this `IsEnabled` runtime check.